git(NavigationNavbarCart): rename component
